# openapi.yaml
openapi: 3.0.3
info:
  title: Secure API
  version: 0.0.1
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /api/health:
    get:
      summary: Health
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
  /api/auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: JWT token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
  /api/echo:
    get:
      summary: Echo input
      parameters:
        - in: query
          name: q
          schema:
            type: string
      responses:
        '200':
          description: Echoed
          content:
            text/plain:
              schema:
                type: string
  /api/users/me:
    get:
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  username:
                    type: string
                  roles:
                    type: array
                    items:
                      type: string
  /api/users/{id}:
    get:
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User info
  /api/admin/stats:
    post:
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Stats